backend/requirements.txt
discord.py
python-dotenv
openai
firebase-admin
aiosqlite
pillow
pandas
pdfminer.six